tech: Support Linux platforms
The requirement to set Manifest files to run
unit tests on Linux platforms has been dropped
since Swift 5.4.
`autoreleasepool` is an Objective-C construct used for
memory management with ARC. It is not needed in Swift
code running on Linux as Swift does not use Objective-C's
memory management model on non-Apple platforms.